Moved code from StapFactory and VideoFactory to Stringfun class
